requirements.txt
python-pptx>=0.6.21
pyyaml>=6.0
